requirements.txt
google-genai
openai
pydantic
ddgs
python-dateutil
dateparser
requests
python-dotenv
rich
